Split jacobian (hessian) functions into jacobian.h
