+++ drake/math/jacobian.cc
@@ -1,0 +1,4 @@
+// For now, this is an empty .cc file that only serves to confirm
+// jacobian.h is a stand-alone header.
+
+#include "drake/math/jacobian.h"
