+++ drake/systems/plants/test/rigid_body_tree/rigid_body_tree_inverse_dynamics_test.cc
@@ -6,6 +6,7 @@
 #include "drake/common/eigen_types.h"
 #include "drake/math/autodiff.h"
 #include "drake/math/autodiff_gradient.h"
+#include "drake/math/jacobian.h"
 #include "drake/systems/plants/joints/floating_base_types.h"
 #include "drake/systems/plants/parser_urdf.h"
